Refactor Monolith worker to integrate with dynamo-run
