--- launch/dynamo-run/engines/llmapi/trtllm_engine.py
@@ -1,9 +1,0 @@
-from llmapi.base_engine import BaseTensorrtLLMEngine, TensorrtLLMEngineConfig
-
-class TensorrtLLMEngine(BaseTensorrtLLMEngine):
-    """
-    Request handler for the generate endpoint
-    """
-
-    def __init__(self, trt_llm_engine_config: TensorrtLLMEngineConfig):
-        super().__init__(trt_llm_engine_config)
